pyproject.toml
[project]
name = "dead-inside-backend"
version = "0.1.0"
description = "At the end of the day, the day's gonna end."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "faster-whisper>=1.1.1",
    "google-cloud-speech>=2.33.0",
    "openai>=1.97.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
    "whisperx>=3.4.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "rich>=13.7.0",
]
